+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 - `Moar.String.underscore/1` is soft-deprecated in favor of `to_case(s, :snake_case)`.
+- Add `Moar.String.to_case(s, :kebab_case)`
 
 ## 1.63.1